+++ src/modules/feature/board/board/Board.vue
@@ -157,6 +157,7 @@ import {
 import { useAppStore, useNotificationStore } from "@data-app";
 import {
 	useBoardAllowedOperations,
+	useBoardCardNavigation,
 	useBoardInactivity,
 	useBoardStore,
 	useCardStore,
@@ -177,7 +178,7 @@ import { SortableEvent } from "sortablejs";
 import { Sortable } from "sortablejs-vue3";
 import { computed, ComputedRef, onMounted, onUnmounted, provide, ref, watch } from "vue";
 import { useI18n } from "vue-i18n";
-import { RouteLocationRaw, useRoute, useRouter } from "vue-router";
+import { useRoute, useRouter } from "vue-router";
 
 const props = defineProps({
 	boardId: { type: String, required: true },
@@ -214,21 +215,7 @@ const cardId = computed(() => {
 	return undefined;
 });
 
-const allCardIds = computed(() => columns.value.flatMap((col) => col.cards.map((c) => c.cardId)));
-
-const currentCardIndex = computed(() => (cardId.value ? allCardIds.value.indexOf(cardId.value) : -1));
-
-const previousCardRoute = computed((): RouteLocationRaw | undefined => {
-	if (currentCardIndex.value <= 0) return undefined;
-	const prevCardId = allCardIds.value[currentCardIndex.value - 1];
-	return { name: "boards-card-detail", params: { boardId: props.boardId, cardId: prevCardId } };
-});
-
-const nextCardRoute = computed((): RouteLocationRaw | undefined => {
-	if (currentCardIndex.value === -1 || currentCardIndex.value >= allCardIds.value.length - 1) return undefined;
-	const nextCardId = allCardIds.value[currentCardIndex.value + 1];
-	return { name: "boards-card-detail", params: { boardId: props.boardId, cardId: nextCardId } };
-});
+const { previousCardRoute, nextCardRoute } = useBoardCardNavigation(computed(() => props.boardId));
 
 const isEditableChipVisible = computed(() => board.value?.readersCanEdit ?? false);
 const hasReadersEditPermission = ref(false);
